Filtrando com seletores
Refatorando também os eventos que são exibidos com o resultado do filtro também no calendário.
