+++ src/state/hooks/useListaEventos.ts
@@ -1,8 +1,8 @@
 import { useRecoilValue } from "recoil";
-import { listaEventosState } from "../atom";
+import { eventosFiltradosState } from "../selectors";
 
 const useListaEventos = () => {
-  return useRecoilValue(listaEventosState);
+  return useRecoilValue(eventosFiltradosState);
 };
 
 export default useListaEventos;
